+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "CHANGELOG.md" file (issue [#22]).
 - CI: Forbid the publication of a version already released (issue [#20]).
 ### Upgraded examples
+- "Calypsonet Terminal Calypso API" to version `1.1.+`
 - "Keyple Service Resource Library" to version `2.0.1`
-- "Keyple Card Calypso Library" to version `2.0.3`
+- "Keyple Card Calypso Library" to version `2.1.0`
 
 ## [2.0.0] - 2021-10-06
 ### Changed